+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adding some annotations from MultiLing 2019
 - Added Dockerized versions of [ROUGE](doc/metrics/docker/rouge.md) and [QAEval](doc/metrics/docker/qaeval.md)
 - Added annotations from [Kryscinski et al. (2019)](doc/datasets/kryscinski2019.md).
+- Added a Dockerized version of [BERTScore](doc/metrics/docker/bertscore.md)
+- Added a Dockerized version of [BLEURT](doc/metrics/docker/bleurt.md)
+- Added a Dockerized version of [MoverScore](doc/metrics/docker/moverscore.md)
 
 ### Changed
 - Moved the actual QAEval metric implementation into the `qaeval` library.